sydani/ats/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==3.0.0
async-timeout==4.0.3
attrs==23.2.0
auth0-python==4.7.2
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
Brotli==1.1.0
build==1.2.2.post1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.23
chromadb-pysqlite3==0.5.3
click==8.1.7
cohere==5.13.3
colorama==0.4.6
coloredlogs==15.0.1
contextlib2==21.6.0
crewai==0.16.3
crewai-tools==0.17.0
cryptography==44.0.0
curl_cffi==0.6.2
dataclasses-json==0.6.4
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docstring-parser==0.15
docx2txt==0.8
duckduckgo_search==4.5.0
durationpy==0.9
email_validator==2.2.0
embedchain==0.1.125
et_xmlfile==2.0.0
exceptiongroup==1.2.0
executing==2.1.0
fastapi==0.110.0
fastavro==1.9.7
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-api-python-client==2.155.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.74.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-generativeai==0.3.2
google-resumable-media==2.7.2
googleapis-common-protos==1.62.0
gptcache==0.1.44
greenlet==3.0.3
grpc-google-iam-v1==0.13.1
grpcio==1.62.0
grpcio-status==1.62.0
grpcio-tools==1.68.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.4
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.27.0
humanfriendly==10.0
hyperframe==6.0.1
idna==3.6
importlib-metadata==6.11.0
importlib_resources==6.4.5
iniconfig==2.0.0
instructor==0.5.2
ipython==8.30.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.2
json_repair==0.31.0
jsonpatch==1.33
jsonpickle==4.0.1
jsonpointer==2.4
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kubernetes==31.0.0
lancedb==0.17.0
langchain==0.1.10
langchain-cohere==0.3.3
langchain-community==0.0.25
langchain-core==0.1.28
langchain-experimental==0.3.3
langchain-google-genai==0.0.9
langchain-openai==0.0.5
langchain-text-splitters==0.0.1
langsmith==0.1.13
litellm==1.55.3
lxml==5.1.0
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.21.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mem0ai==0.1.35
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mutagen==1.47.0
mypy-extensions==1.0.0
networkx==3.4.2
nodeenv==1.9.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.13.3
openpyxl==3.1.5
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-exporter-otlp-proto-http==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
orjson==3.9.15
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.3
parameterized==0.9.0
parso==0.8.4
pdfminer.six==20240706
pdfplumber==0.11.4
pexpect==4.9.0
pillow==11.0.0
pip-review==1.3.0
pluggy==1.5.0
portalocker==3.0.0
posthog==3.7.4
primp==0.9.0
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.23.0
protobuf==4.25.3
ptyprocess==0.7.0
pulsar-client==3.5.0
pure_eval==0.2.3
py==1.11.0
pyarrow==18.1.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.21.0
pydantic==2.6.3
pydantic-extra-types==2.10.1
pydantic-settings==2.7.0
pydantic_core==2.16.3
PyGithub==2.5.0
Pygments==2.17.2
PyJWT==2.10.1
pylance==0.20.0
PyNaCl==1.5.0
pyparsing==3.2.0
pypdf==5.1.0
PyPDF2==3.0.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyright==1.1.390
pysbd==0.3.4
PySocks==1.7.1
pysqlite3-binary==0.5.4
pytest==8.3.4
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-docx-1==0.0.2
python-dotenv==1.0.1
python-multipart==0.0.20
pytube==15.0.0
pytz==2024.2
pyvis==0.3.2
PyYAML==6.0.1
qdrant-client==1.12.1
ratelimiter==1.2.0.post0
referencing==0.35.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retry==0.9.2
rich==13.7.1
rpds-py==0.22.3
rsa==4.9
schema==0.7.7
selenium==4.27.1
semver==3.0.2
setuptools==75.6.0
shapely==2.0.6
shellingham==1.5.4
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.27
stack-data==0.6.3
starlette==0.36.3
sympy==1.13.3
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.5.2
tokenizers==0.21.0
tomli==2.2.1
tomli_w==1.1.0
tqdm==4.66.2
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
typer==0.9.0
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.1
uv==0.5.9
uvicorn==0.27.1
uvloop==0.21.0
watchfiles==0.21.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
youtube-transcript-api==0.6.3
yt-dlp==2024.12.13
zipp==3.17.0
